+++ language/english.php
@@ -89,7 +89,7 @@ $LANG_AC = array (
     // rrule
 
     'repeat'            => 'Repeat',
-    'none'              => 'None',
+    'none'              => 'No Repeat',
     'hourly'            => 'Hourly',
     'daily'             => 'Daily',
     'weekly'            => 'Weekly',
@@ -147,6 +147,7 @@ $LANG_AC_JS = array(
     'err_end_before_start' => 'End date/time must be greater than start date/time',
     'err_initialize'    => 'Error Initializing Agenda Plugin',
     'event_queued'      => 'Thank you for your submission. Your event submission has been placed in the queue for review and approval.',
+    'exception_event'   => 'Exception Event',
 );
 
 $LANG_AC_ERRORS = array(