+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Agenda Plugin ChangeLog
 
 ## v0.3.1
+ - Disable Edit Series button on repeating events that are marked as an exception
  - Left align buttons on both pop-up and edit strings
  - Reduced the size of the description entry field by 1 row
  - Fixed some E_ALL errors